add auth to posts mutations

+++ prisma/src/resolvers/Mutation/post.mutations.js
@@ -6,10 +6,24 @@ export const createPost = (parent, { data: { title, body, published } }, { prism
   return prisma.mutation.createPost({ data }, info);
 };
 
-export const updatePost = (parent, { id, data }, { prisma }, info) => {
+export const updatePost = async (_, { id, data }, { prisma, request }, info) => {
+  const userId = getUserId(request);
+  const postExists = await prisma.exists.Post({ id, author: { id: userId } });
+
+  if (!postExists) {
+    throw new Error('Post not found');
+  }
+
   return prisma.mutation.updatePost({ data, where: { id } }, info);
 };
 
-export const deletePost = (parent, { id }, { prisma }, info) => {
+export const deletePost = async (parent, { id }, { prisma, request }, info) => {
+  const userId = getUserId(request);
+  const postExists = await prisma.exists.Post({ id, author: { id: userId } });
+
+  if (!postExists) {
+    throw new Error('Post not found');
+  }
+
   return prisma.mutation.deleteUser({ where: { id } }, info);
 };
